Create applied_patches folder + add patch files used.
The NimBLE core library that this project is based on has been patched
to improve functionality. This new folder contains the patch files applied
that have not yet been merged offically into the esp-nimble fork.
The offical-merged folder contains patches that are offically applied to the upstream
NimBLE core, the un-merged folder contains the patches not yet merged anywhere.